third_party/blink/web_tests/platform/generic/fast/webgl/canvas-to-data-url-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

Check if back buffer (toDataURL) is equal to front buffer (getImageSourceURL)
1) when drawingBuffer is preserved.
PASS preserve_canvas3D.toDataURL('image/png') == internals.getImageSourceURL(preserve_canvas3D) is true
2) when drawingBuffer is not preserved.
PASS nonpreserve_canvas3D.toDataURL('image/png') == internals.getImageSourceURL(nonpreserve_canvas3D) is true
Check if back buffer (toDataURL) is equal to front buffer (getImageSourceURL) one frame after drawing webgl contents.
1) when drawingBuffer is preserved.
PASS preserve_canvas3D.toDataURL('image/png') == internals.getImageSourceURL(preserve_canvas3D) is true
2) when drawingBuffer is not preserved. They must be different.
PASS nonpreserve_canvas3D.toDataURL('image/png') != internals.getImageSourceURL(nonpreserve_canvas3D) is true
 
